changed docker-compose,yml to run notification client in interactive mode, moved notification publish class to its own library, removed rabbitmq dependency from API
