+++ Digital_Excellence.sln
@@ -24,7 +24,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		.editorconfig = .editorconfig
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NotificationSystem", "NotificationSystem\NotificationSystem.csproj", "{A0F1BB66-1749-47B7-940F-2EA6D004F33B}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NotificationSystem", "NotificationSystem\NotificationSystem.csproj", "{A0F1BB66-1749-47B7-940F-2EA6D004F33B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MessagebrokerPublisher", "MessagebrokerPublisher\MessagebrokerPublisher.csproj", "{3A0B6EED-3F6B-43F7-9C1E-0311EC9871CB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -68,6 +70,10 @@ Global
 		{A0F1BB66-1749-47B7-940F-2EA6D004F33B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A0F1BB66-1749-47B7-940F-2EA6D004F33B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A0F1BB66-1749-47B7-940F-2EA6D004F33B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3A0B6EED-3F6B-43F7-9C1E-0311EC9871CB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3A0B6EED-3F6B-43F7-9C1E-0311EC9871CB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3A0B6EED-3F6B-43F7-9C1E-0311EC9871CB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3A0B6EED-3F6B-43F7-9C1E-0311EC9871CB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
